Refactoring: Separated some stuff from main file
